fix up cron
